fix maxlength, fontScaling, validation
