+++ src/components/transaction-sum/index.ts
@@ -1,0 +1,1 @@
+export * from './transaction-sum';
